+++ CHANGELOG.md
@@ -11,7 +11,12 @@ All notable changes to this project will be documented in this file, following t
 - Breaking: Change multiple geometrical primitive parameters
 - Multi-state data model tweaks
 - Support for representation-specific parameters (customize presentation visuals)
+- Add `coarse` component kind
+- Add `spacefill` representation
+- Add `element_granularity` field to component expressions
+- Add I/HM basic restraints example
 - Add primitives: ellipse, ellipsoid, box, arrow
+- Add basic support for volumetric data
 - Add membrane orientation examples